Add GPF prefix to the Google Shopping
GPF prefix added to the Google Shopping section. This should be configured from admin in the future: an optional prefix.

Almost all Google shopping product feeds will carry a prefix in the ID, if the code is generated by Google Product Feeds for WooCommerce, and the feed "name" is woocommerce, it is in form "woocommerce_gpf_PRODUCTID" - hence this is hard-wired in to check validation.

When validation is confirmed, extra coding will be pushed which will add this as an admin extra field, to allow GUI on this element.